Add different points for each ball and a multiplier for the target, the random generation of all of it and integrate it

+++ include/core/DifficultyManager.hpp
@@ -63,7 +63,7 @@ private:
 	sf::Font font_;
 	sf::Text timer_;
 
-	int score_;
+	float score_;
 	int objective_;
 	int objectiveIncrement_;
 	sf::Text scoreText_;
